% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{Arc_term}
\alias{Arc_term}
\title{The arc effect subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
Arc_term(Rrup, Rhypo, faba, Ftype, V)
}
\description{
This subroutine function of BCHydro 2011 calculates Forearc/Backarc scaling term.
DO NOT Run it. See bchydro_2011() for prediction.
}
