% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_2014_nga.R
\name{ask_2014_nga}
\alias{ask_2014_nga}
\title{The GMPE of ASK 2014}
\usage{
ask_2014_nga(
  M,
  T = 1000,
  CRjb = 999.9,
  Rrup,
  Rjb,
  Rx,
  Ry0 = 999,
  Ztor = 999,
  dip,
  lambda,
  after_shock = 0,
  HW,
  W = 999,
  Z10 = 999,
  Vs30,
  Vs30_code,
  region
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{T}{Period (sec); Use Period = -1 for PGV computation.
Use 1000 (by default) for output the array of Sa with original NGA West2 periods}

\item{CRjb}{Centroid CRjb, assumed to be 999.9 here -> assume no aftershock}

\item{Rrup}{Closest distance (km) to the ruptured plane}

\item{Rjb}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{Rx}{Site coordinate (km) measured perpendicular to the fault strike
from the fault line with down-dip direction to be positive}

\item{Ry0}{Horizontal distance off the end of the rupture measured parallel to strike. 999 if unknown}

\item{Ztor}{Depth(km) to the top of ruptured plane. 999 if unknown}

\item{dip}{Fault dip angle (in degree)}

\item{lambda}{Rake angle (in degree)}

\item{after_shock}{Flag for aftershocks. 0 for Class 1 (no aftershock), 1 for Class 2.}

\item{HW}{Falg for hanging wall sites. 1 for sites on the hanging wall side of the fault, 0 otherwise.}

\item{W}{Down-dip rupture width (km). 999 if unknown}

\item{Z10}{Basin depth (km): depth from the groundsurface to the 1km/s shear-wave horizon.
999 if unknown.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s). Reference site condition is
1130 m/s.}

\item{Vs30_code}{Code for Vs30 measurement. 0 for measured Vs30; 1 for inferred Vs30}

\item{region}{Region indicator: 0 for global; 1 for California; 2 for Japan; 3 for China;
4 for Italy; 5 for Turkey; 6 for Taiwan.}
}
\value{
A list of five elements is returned: med - median spectral acceleration prediction (in g);
sigma - logarithmic standard deviation of spectral acceleration prediction; phi - logarithmic
standard deviation of within event residuals; tau - logarithmic standard deviation of between
event residuals; period - the corresponding oscillator periods
}
\description{
This function calculates the ground motion median values and standard deviations
}
\examples{
ask_2014_nga(M = 5, T = 1000, Rrup = 90, Rjb = 85, Rx = 85, Ry0 = 80, dip = 80,
lambda = 75, after_shock = 0, HW = 1, W = 10, region = 1, Vs30 = 350, Vs30_code = 0)
}
\references{
Abrahamson, N. A., Silva, W. J., and Kamai, R. (2014). Summary of the
ASK14 Ground Motion Relation for Active Crustal Regions. Earthquake Spectra, 30(3),
1025-1055.
}
