% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_2014_nga.R
\name{ask_2014_subroutine}
\alias{ask_2014_subroutine}
\title{The subroutine of GMPE of ASK 2014}
\usage{
ask_2014_subroutine(
  M,
  ip,
  R_RUP,
  R_JB,
  Rx,
  Ry0 = 999,
  Ztor = 999,
  dip,
  F_RV,
  F_NM,
  after_shock,
  HW,
  W,
  Z10 = 999,
  Vs30,
  Vs30_code,
  region,
  CRjb = 999.9
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{ip}{The index of working period in the per-defined periods: (0.01, 0.02,    0.03,
0.05,    0.075,    0.1,    0.15,    0.2,    0.25,    0.3,    0.4,    0.5,    0.75,1,    1.5,    2,    3,    4,    5,
6,    7.5,    10, 0, -1)}

\item{R_RUP}{Closest distance (km) to the ruptured plane}

\item{R_JB}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{Rx}{Site coordinate (km) measured perpendicular to the fault strike
from the fault line with down-dip direction to be positive}

\item{Ry0}{Horizontal distance off the end of the rupture measured parallel to strike. 999 if unknown}

\item{Ztor}{Depth(km) to the top of ruptured plane. 999 if unknown}

\item{dip}{Fault dip angle (in degree)}

\item{F_RV}{Flag of fault type. 1 for Reverse; 0 otherwise}

\item{F_NM}{Flag of fault type. 1 for Normal; 0 otherwise}

\item{after_shock}{Flag for aftershocks. 0 for Class 1 (no aftershock), 1 for Class 2.}

\item{HW}{Falg for hanging wall sites. 1 for sites on the hanging wall side of the fault, 0 otherwise.}

\item{W}{Down-dip rupture width (km). 999 if unknown}

\item{Z10}{Basin depth (km): depth from the groundsurface to the 1km/s shear-wave horizon.
999 if unknown.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s). Reference site condition is
1130 m/s.}

\item{Vs30_code}{Code for Vs30 measurement. 0 for measured Vs30; 1 for inferred Vs30}

\item{region}{Region indicator: 0 for global; 1 for California; 2 for Japan; 3 for China;
4 for Italy; 5 for Turkey; 6 for Taiwan.}

\item{CRjb}{Centroid CRjb, assumed to be 999.9 here -> assume no aftershock}
}
\value{
An array of four values: (1) median spectral acceleration prediction; (2)
logarithmic standard deviation of spectral acceleration prediction; (3) logarithmic
standard deviation of within event residuals; and (4) logarithmic standard deviation of between
event residuals at the period of ip.
}
\description{
This is a subroutine of ASK 2014
}
\references{
Abrahamson, N. A., Silva, W. J., and Kamai, R. (2014). Summary of the
ASK14 Ground Motion Relation for Active Crustal Regions. Earthquake Spectra, 30(3),
1025-1055.
}
