% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{BCHHR2Vs760}
\alias{BCHHR2Vs760}
\title{The amplifcation subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
BCHHR2Vs760(T_760, LSa_noamp)
}
\description{
This subroutine function of BCHydro 2011 calculates amplifcation factors.
DO NOT Run it. See bchydro_2011() for prediction.
}
