% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{bchydro_2011}
\alias{bchydro_2011}
\title{The GMPE of BCHydro 2011 for subduction earthquakes}
\usage{
bchydro_2011(
  M,
  T = 1000,
  Vs30,
  Rrup,
  Rhypo,
  Ftype,
  faba,
  depth,
  flag_deltaC1 = 1
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{T}{Period (sec) (range should be between 0 s and 10 s).
Use 1000 (by default) for output the array of Sa with original BCHyrdo per-defined periods}

\item{Vs30}{The time averaged shear wave velocity on the top 30 m.}

\item{Rrup}{Closest distance (km) to the ruptured plane}

\item{Rhypo}{Hypocentral distance(km)}

\item{Ftype}{Flag for interface event or intraslab events:
0 for Interface - use rupture distance; 1 for Intraslab - use hypocentral distance.}

\item{faba}{Flag for BackArc sites: 0 for Non-Backarc sites; 1 for Backarc sites}

\item{depth}{The depth of hypocenter (km)}

\item{flag_deltaC1}{Flag for Recommended deltaC1 Values:
0 for Recommended deltaC1 Values NOT period dependent
(-0.5, 0.0, 0.5) as on the 06/2010 model;
1 for Recommended deltaC1 Values Period dependent as on the 08/2011 model.
Note: here we implement only the CENTRAL deltaC1 values!}
}
\value{
A list of four elements is returned: Sa - median spectral acceleration prediction (in g);
RockSa - median spectral acceleration prediction on Rock reference site (in g);
sigma - totla standard deviation (log); period - the corresponding oscillator periods
}
\description{
This function calculates the ground motion median values and standard deviations
}
\examples{
bchydro_2011(M = 6, T = 1000, Vs30 = 450, Rrup = 85, Rhypo = 100,
Ftype = 0, faba = 0, depth = 25, flag_deltaC1 = 1)

bchydro_2011(M = 7.5, T = c(0.01, 2.5), Vs30 = 450, Rrup = 85, Rhypo = 100,
Ftype = 1, faba = 1, depth = 65, flag_deltaC1 = 1)
}
\references{
BC Hydro subduction Ground motion model.
A report to BC Hydro Engineering No. E658 - Vol. 3 May 2009 by Abrahamson, et al.
}
