% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{BCHydro_2011_subroutine}
\alias{BCHydro_2011_subroutine}
\title{The subroutine of BCHydro 2011 for subduction earthquakes}
\usage{
BCHydro_2011_subroutine(
  M,
  Vs30,
  T,
  Rrup,
  Rhypo,
  Ftype,
  faba,
  depth,
  flag_deltaC1,
  PGA_rock
)
}
\description{
This is a subroutine function of BCHydro 2011. Run bchydro_2011() for prediction.
}
