% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{BCHydro_sigma}
\alias{BCHydro_sigma}
\title{The sigma subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
BCHydro_sigma(V)
}
\description{
This subroutine function of BCHydro 2011 estimates model uncertainty (standard deviation)
terms. DO NOT Run it. See bchydro_2011() for prediction.
}
