% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssa_2014_coeffs.R
\docType{data}
\name{bssa_2014_coeffs}
\alias{bssa_2014_coeffs}
\title{The coefficients of BSSA 2014}
\format{
A data frame with 107 rows (number of oscillator) and 37 variables
}
\source{
\url{https://journals.sagepub.com/doi/suppl/10.1193/070113EQS184M}
}
\usage{
bssa_2014_coeffs
}
\description{
A dataset containing all coefficients for BSSA 2014
}
\keyword{datasets}
