% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssa_2014_nga.R
\name{bssa_2014_nga}
\alias{bssa_2014_nga}
\title{The GMPE of BSSA 2014}
\usage{
bssa_2014_nga(
  M,
  T = 1000,
  Rjb,
  Fault_Type,
  region,
  z1 = 999,
  Vs30,
  return.type = 1
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{T}{Period (sec); Use Period = -1 for PGV computation.
Use 1000 (by default) for output the array of Sa with original NGA West2 periods}

\item{Rjb}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{Fault_Type}{The indicator of fault type: 0 for unspecified fault;
1 for strike-slip fault; 2 for normal fault; 3 for reverse fault}

\item{region}{Region indicator: 0 for global (incl. Taiwan); 1 for California;
2 for Japan; 3 for China or Turkey; 4 for Italy}

\item{z1}{Basin depth (km): depth from the groundsurface to the 1km/s shear-wave horizon.
999 if unknown.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s)}

\item{return.type}{The indicator specifies which type of data return:
1 for the med (in g)/sigma/phi/tau/period; 2 for F_E/F_P/F_S;
3 for r/Ln_Flin/Ln_Fnlin/f2/F_dz1/PGAr}
}
\value{
See return.type for the returned results
}
\description{
This function calculates the ground motion median values and standard deviations
}
\examples{
bssa_2014_nga(M = 5, T = 1000, Rjb = 85, Fault_Type = 1, region = 1, z1 = 999, Vs30 = 350)
}
\references{
Boore, D. M., Stewart, J. P., Seyhan, E., and Atkinson, G. M. (2014).
NGA-West2 Equations for Predicting PGA, PGV, and 5% Damped PSA for
Shallow Crustal Earthquakes. Earthquake Spectra, 30(3), 1057-1085.
}
