% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssa_2014_nga.R
\name{bssa_2014_subroutine}
\alias{bssa_2014_subroutine}
\title{The subroutine of GMPE of BSSA 2014}
\usage{
bssa_2014_subroutine(
  M,
  ip,
  Rjb,
  U,
  SS,
  NS,
  RS,
  region,
  z1,
  Vs30,
  return.type = 1,
  coeffs = bssa_2014_coeffs
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{ip}{The index of working period in the per-defined periods: (-1,
0, 0.010, 0.020, 0.022, 0.025, 0.029, 0.030, 0.032, 0.035, 0.036, 0.040, 0.042,
0.044, 0.045, 0.046, 0.048, 0.050, 0.055, 0.060, 0.065, 0.067, 0.070, 0.075,
0.080, 0.085, 0.090, 0.095, 0.100, 0.110, 0.120, 0.130, 0.133, 0.140, 0.150,
0.160, 0.170, 0.180, 0.190, 0.200, 0.220, 0.240, 0.250, 0.260, 0.280, 0.290,
0.300, 0.320, 0.340, 0.350, 0.360, 0.380, 0.400, 0.420, 0.440, 0.450, 0.460,
0.480, 0.500, 0.550, 0.600, 0.650, 0.667, 0.700, 0.750, 0.800, 0.850, 0.900,
0.950, 1.000, 1.100, 1.200, 1.300, 1.400, 1.500, 1.600, 1.700, 1.800, 1.900,
2.000, 2.200, 2.400, 2.500, 2.600, 2.800, 3.000, 3.200, 3.400, 3.500, 3.600,
3.800, 4.000, 4.200, 4.400, 4.600, 4.800, 5.000, 5.500, 6.000, 6.500, 7.000,
7.500, 8.000, 8.500, 9.000, 9.500, 10.000)}

\item{Rjb}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{U}{The indicator of fault type: 1 for unspecified fault, 0 for otherwise}

\item{SS}{The indicator of fault type: 1 for strike-slip fault, 0 for otherwise}

\item{NS}{The indicator of fault type: 1 for normal fault, 0 for otherwise}

\item{RS}{The indicator of fault type: 1 for reverse fault, 0 for otherwise}

\item{region}{Region indicator: 0 for global (incl. Taiwan); 1 for California;
2 for Japan; 3 for China or Turkey; 4 for Italy}

\item{z1}{Basin depth (km): depth from the groundsurface to the 1km/s shear-wave horizon.
999 if unknown.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s)}

\item{return.type}{The indicator specifies which type of data return:
1 for the med/sigma/phi/tau; 2 for F_E/F_P/F_S;
3 for r/Ln_Flin/Ln_Fnlin/f2/F_dz1/PGAr}

\item{coeffs}{The coefficient table of BSSA 2014. You can use the internal saved data object, bssa_2014_coeffs}
}
\value{
See return.type for the returned results
}
\description{
This is a subroutine of BSSA 2014
}
\references{
Boore, D. M., Stewart, J. P., Seyhan, E., and Atkinson, G. M. (2014).
NGA-West2 Equations for Predicting PGA, PGV, and 5% Damped PSA for
Shallow Crustal Earthquakes. Earthquake Spectra, 30(3), 1057-1085.
}
