% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_subreg.R
\docType{data}
\name{CA_subreg}
\alias{CA_subreg}
\title{The polygons for CA subregional anelastic attenuation Model}
\format{
A data frame with 10 polygon featrues
}
\source{
\url{https://doi.org/10.1785/0120220173}
}
\usage{
CA_subreg
}
\description{
A dataset containing polygons of subregions for Buckreis et al (2023) model
}
\keyword{datasets}
