% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_subreg_coeffs.R
\docType{data}
\name{CA_subreg_coeffs}
\alias{CA_subreg_coeffs}
\title{The coefficients of Buckreis et al 2023 CA subregional anelastic attenuation model}
\format{
A data frame with 107 rows (number of oscillator) and 22 variables
}
\source{
\url{https://doi.org/10.1785/0120220173}
}
\usage{
CA_subreg_coeffs
}
\description{
A dataset containing all coefficients for Buckreis et al 2023
}
\references{
Buckreis, T.E., J.P. Stewart, S.J. Brandenberg, and P. Wang (2023) Subregional anelastic attenuation model for California, Bulletin of the Seimological Society of America. 2023. doi:10.1785/0120220173
}
\keyword{datasets}
