% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssa_2014_nga.R
\name{CA_subreg_path_calc}
\alias{CA_subreg_path_calc}
\title{The subroutine of CA subregional path correction}
\usage{
CA_subreg_path_calc(site_lon, site_lat, clst_lon, clst_lat)
}
\arguments{
\item{site_lon}{The longitude of site (in WGS84 degree).}

\item{site_lat}{The latitude of site (in WGS84 degree).}

\item{clst_lon}{Longitude (in WGS84 degrees) of closest point on the surface projection of the rupture surface to the site.}

\item{clst_lat}{Latitude (in WGS84 degrees) of closest point on the surface projection of the rupture surface to the site.}
}
\value{
A list of three elements: 1) the site-to-source distance;
2) the distances of the path within each CA subregion; and 3) the CA subregion index where the source is.
}
\description{
This is a subroutine of CA subregional path correction
}
\examples{
CA_subreg_path_calc(site_lon = -118.8713, site_lat = 35.80407, clst_lon = -116.9274, clst_lat = 33.6204)
}
