% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{calc_val}
\alias{calc_val}
\title{The subroutine function of BCHydro 2011 for PGA for subduction earthquakes}
\usage{
calc_val(M, Rrup, Rhypo, PGA_rock, Vs30, Ftype, faba, depth, V)
}
\description{
This subroutine function of BCHydro 2011 predicts for PGA.
DO NOT Run it. See bchydro_2011() for prediction.
}
