% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_2014_nga.R
\name{cb_2014_nga}
\alias{cb_2014_nga}
\title{The GMPE of CB 2014}
\usage{
cb_2014_nga(
  M,
  T = 1000,
  Rrup,
  Rjb,
  Rx,
  W = 999,
  Ztor = 999,
  Zbot,
  dip,
  lambda,
  Fhw,
  Vs30,
  Z25 = 999,
  Zhyp = 999,
  region
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{T}{Period (sec); Use Period = -1 for PGV computation.
Use 1000 (by default) for output the array of Sa with original NGA West2 periods}

\item{Rrup}{Closest distance (km) to the ruptured plane}

\item{Rjb}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{Rx}{Site coordinate (km) measured perpendicular to the fault strike
from the fault line with down-dip direction to be positive}

\item{W}{Down-dip rupture width (km). 999 if unknown}

\item{Ztor}{Depth(km) to the top of ruptured plane. 999 if unknown}

\item{Zbot}{Depth(km) to the bottom of the seismogenic crust. It is needed only when W is unknown.}

\item{dip}{Fault dip angle (in degree)}

\item{lambda}{Rake angle (in degree)}

\item{Fhw}{Falg for hanging wall sites. 1 for sites on the hanging wall side of the fault, 0 otherwise.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s)}

\item{Z25}{Basin depth (km): depth from the groundsurface to the 2.5km/s shear-wave horizon.
999 if unknown.}

\item{Zhyp}{Hypocentral depth of the earthquake measured from sea level. 999 in unknown}

\item{region}{Region indicator: 0 for global (incl. Taiwan); 1 for California; 2 for Japan;
3 for China or Turkey; 4 for Italy}
}
\value{
A list of five elements is returned: med - median spectral acceleration prediction (in g);
sigma - logarithmic standard deviation of spectral acceleration prediction; phi - logarithmic
standard deviation of within event residuals; tau - logarithmic standard deviation of between
event residuals; period - the corresponding oscillator periods
}
\description{
This function calculates the ground motion median values and standard deviations
}
\examples{
cb_2014_nga(M = 5, T = 1000, Rrup = 90, Rjb = 85, Rx = 85, W = 10, Zbot = 15, dip = 80,
lambda = 75, Fhw = 1, Vs30 = 350, region = 1, Zhyp = 8, Z25 = 999)
}
\references{
Campbell, K. W., and Bozorgnia, Y. (2014). NGA-West2 Ground Motion Model
for the Average Horizontal Components of PGA, PGV, and 5% Damped Linear
Acceleration Response Spectra. Earthquake Spectra, 30(3), 1087-1115.
}
