% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_2014_nga.R
\name{cb_2014_subroutine}
\alias{cb_2014_subroutine}
\title{The subroutine of GMPE of CB 2014}
\usage{
cb_2014_subroutine(
  M,
  ip,
  Rrup,
  Rjb,
  Rx,
  W,
  Ztor,
  Zbot,
  dip,
  Fhw,
  Vs30,
  Z25in,
  Zhyp,
  lambda,
  Frv,
  Fnm,
  region,
  A1100 = 999
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{ip}{The index of working period in the per-defined periods: (0.010, 0.020, 0.030, 0.050, 0.075,
0.10, 0.15, 0.20, 0.25, 0.30, 0.40, 0.50, 0.75, 1.0, 1.5, 2.0, 3.0, 4.0, 5.0, 7.5, 10.0, 0, -1)}

\item{Rrup}{Closest distance (km) to the ruptured plane}

\item{Rjb}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{Rx}{Site coordinate (km) measured perpendicular to the fault strike
from the fault line with down-dip direction to be positive}

\item{W}{Down-dip rupture width (km). 999 if unknown}

\item{Ztor}{Depth(km) to the top of ruptured plane. 999 if unknown}

\item{Zbot}{Depth(km) to the bottom of the seismogenic crust. It is needed only when W is unknown.}

\item{dip}{Fault dip angle (in degree)}

\item{Fhw}{Falg for hanging wall sites. 1 for sites on the hanging wall side of the fault, 0 otherwise.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s)}

\item{Z25in}{Basin depth (km): depth from the groundsurface to the 2.5km/s shear-wave horizon.
999 if unknown.}

\item{Zhyp}{Hypocentral depth of the earthquake measured from sea level. 999 if unknown.}

\item{lambda}{Rake angle (in degree)}

\item{Frv}{Flag of reverse fault. 1 for reverse fault, 0 otherwise.}

\item{Fnm}{Flag of normal fault. 1 for normal fault, 0 otherwise.}

\item{region}{Region indicator: 0 for global (incl. Taiwan); 1 for California; 2 for Japan;
3 for China or Turkey; 4 for Italy}

\item{A1100}{Median estimated value of PGA on rock with Vs30 = 1100m/s. 999 if unknown}
}
\value{
An array of four values: (1) median spectral acceleration prediction; (2)
logarithmic standard deviation of spectral acceleration prediction; (3) logarithmic
standard deviation of within event residuals; and (4) logarithmic standard deviation of between
event residuals at the period of ip.
}
\description{
This is a subroutine of CB 2014
}
\references{
Campbell, K. W., and Bozorgnia, Y. (2014). NGA-West2 Ground Motion Model
for the Average Horizontal Components of PGA, PGV, and 5% Damped Linear
Acceleration Response Spectra. Earthquake Spectra, 30(3), 1087-1115.
}
