% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy_2014_nga.R
\name{cy_2014_nga}
\alias{cy_2014_nga}
\title{The GMPE of CY 2014}
\usage{
cy_2014_nga(
  M,
  T = 1000,
  Rrup,
  Rjb,
  Rx,
  Ztor,
  dip,
  lambda,
  Z10 = 999,
  Vs30,
  Vs30_code,
  Fhw,
  region,
  d_DPP = 0
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{T}{Period (sec); Use Period = -1 for PGV computation.
Use 1000 (by default) for output the array of Sa with original NGA West2 periods}

\item{Rrup}{Closest distance (km) to the ruptured plane}

\item{Rjb}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{Rx}{Site coordinate (km) measured perpendicular to the fault strike
from the fault line with down-dip direction to be positive}

\item{Ztor}{Depth(km) to the top of ruptured plane. 999 if unknown}

\item{dip}{Fault dip angle (in degree)}

\item{lambda}{Rake angle (in degree)}

\item{Z10}{Basin depth (km): depth from the groundsurface to the 1km/s shear-wave horizon.
999 if unknown.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s). Reference site condition is
1130 m/s.}

\item{Vs30_code}{Code for Vs30 measurement. 0 for measured Vs30; 1 for inferred Vs30}

\item{Fhw}{Falg for hanging wall sites. 1 for sites on the hanging wall side of the fault, 0 otherwise.}

\item{region}{Region indicator: 0 for global (incl. Taiwan); 1 for California; 2 for Japan;
3 for China; 4 for Italy; 5 for Turkey}

\item{d_DPP}{Centered on the site and earthquake specific average. d_DPP = 0 for median calc (by default)}
}
\value{
A list of five elements is returned: med - median spectral acceleration prediction;
sigma - logarithmic standard deviation of spectral acceleration prediction; phi - logarithmic
standard deviation of within event residuals; tau - logarithmic standard deviation of between
event residuals; period - the corresponding oscillator periods
}
\description{
This function calculates the ground motion median values and standard deviations
}
\examples{
cb_2014_nga(M = 5, T = 1000, Rrup = 90, Rjb = 85, Rx = 85, W = 10, Zbot = 15, dip = 80,
lambda = 75, Fhw = 1, Vs30 = 350, region = 1, Zhyp = 8, Z25 = 999)
}
\references{
Chiou, B. S.-J., and Youngs, R. R. (2014). Update of the Chiou and
Youngs NGA Model for the Average Horizontal Component of Peak Ground
Motion and Response Spectra. Earthquake Spectra, 30(3), 1117-1153.
}
