% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy_2014_nga.R
\name{cy_2014_subroutine}
\alias{cy_2014_subroutine}
\title{The subroutine of GMPE of CY 2014}
\usage{
cy_2014_subroutine(
  M,
  ip,
  R_RUP,
  R_JB,
  Rx,
  Ztor,
  dip,
  F_RV,
  F_NM,
  HW,
  Z10,
  Vs30,
  Vs30_code,
  region,
  d_DPP
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{ip}{The index of working period in the per-defined periods: (-1, 0, 0.01,    0.02,    0.03,    0.04,    0.05,
0.075,    0.1,    0.12,    0.15,    0.17,    0.2,    0.25,    0.3, 0.4,    0.5,    0.75,    1,    1.5,    2,    3,    4,    5,    7.5,    10)}

\item{R_RUP}{Closest distance (km) to the ruptured plane}

\item{R_JB}{Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane}

\item{Rx}{Site coordinate (km) measured perpendicular to the fault strike
from the fault line with down-dip direction to be positive}

\item{Ztor}{Depth(km) to the top of ruptured plane. 999 if unknown}

\item{dip}{Fault dip angle (in degree)}

\item{F_RV}{Flag of reverse fault. 1 for reverse fault, 0 otherwise.}

\item{F_NM}{Flag of normal fault. 1 for normal fault, 0 otherwise.}

\item{HW}{Falg for hanging wall sites. 1 for sites on the hanging wall side of the fault, 0 otherwise.}

\item{Z10}{Basin depth (km): depth from the groundsurface to the 1km/s shear-wave horizon.
999 if unknown.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s). Reference site condition is
1130 m/s.}

\item{Vs30_code}{Code for Vs30 measurement. 0 for measured Vs30; 1 for inferred Vs30}

\item{region}{Region indicator: 0 for global (incl. Taiwan); 1 for California; 2 for Japan;
3 for China; 4 for Italy; 5 for Turkey}

\item{d_DPP}{Centered on the site and earthquake specific average. d_DPP = 0 for median calc (by default)}
}
\value{
An array of four values: (1) median spectral acceleration prediction; (2)
logarithmic standard deviation of spectral acceleration prediction; (3) logarithmic
standard deviation of within event residuals; and (4) logarithmic standard deviation of between
event residuals at the period of ip.
}
\description{
This is a subroutine of CY 2014
}
\references{
Chiou, B. S.-J., and Youngs, R. R. (2014). Update of the Chiou and
Youngs NGA Model for the Average Horizontal Component of Peak Ground
Motion and Response Spectra. Earthquake Spectra, 30(3), 1117-1153.
}
