% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{D_term}
\alias{D_term}
\title{The distance subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
D_term(M, Rrup, Rhypo, Ftype, V)
}
\description{
This subroutine function of BCHydro 2011 calculates distance scaling base model
with R term - either Rupture Distance (Interface events) of Hypocentral distance (Intraslab events).
DO NOT Run it. See bchydro_2011() for prediction.
}
