% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{Depth_term}
\alias{Depth_term}
\title{The depth subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
Depth_term(depth, Ftype, V)
}
\description{
This subroutine function of BCHydro 2011 calculates depth scaling term.
DO NOT Run it. See bchydro_2011() for prediction.
}
