% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{get_BCHydro_constants}
\alias{get_BCHydro_constants}
\title{The constant terms subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
get_BCHydro_constants(index, flag_deltaC1)
}
\description{
This subroutine function of BCHydro 2011 gets constant terms.
DO NOT Run it. See bchydro_2011() for prediction.
}
