% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{get_BCHydro_constants_760}
\alias{get_BCHydro_constants_760}
\title{The constant terms subroutine function of BCHydro 2011 for Vs760 for subduction earthquakes}
\usage{
get_BCHydro_constants_760(index_760)
}
\description{
This subroutine function of BCHydro 2011 gets constant term for the reference site with
Vs30 = 760 m/s. DO NOT Run it. See bchydro_2011() for prediction.
}
