% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hea_2020_cena_fnl_coeffs.R
\docType{data}
\name{hea_2020_cena_fnl_coeffs}
\alias{hea_2020_cena_fnl_coeffs}
\title{The coefficients of Hashash et al 2020 CENA Fnl Coefficients}
\format{
A data frame with 40 rows (number of oscillator) and 7 variables
}
\usage{
hea_2020_cena_fnl_coeffs
}
\description{
A dataset containing all coefficients for Hashash et al 2020 CENA Fnl
}
\references{
Hashash YMA, Ilhan O, Harmon JA, et al. Nonlinear site amplification model for ergodic seismic hazard analysis in Central and Eastern North America. Earthquake Spectra. 2020;36(1):69-86. doi:10.1177/8755293019878193
}
\keyword{datasets}
