% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_2014_nga.R
\name{i_2014_subroutine}
\alias{i_2014_subroutine}
\title{The subroutine of GMPE of Idriss 2014}
\usage{
i_2014_subroutine(M, ip, R_RUP, Vs30, Fault_Type)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{ip}{The index of working period in the per-defined periods: (0.01,    0.02,    0.03,    0.05,    0.075,    0.1,
0.15,    0.2,    0.25,    0.3,    0.4,    0.5,    0.75, 1,    1.5,    2,    3,    4,    5,    7.5,    10, 0)}

\item{R_RUP}{Closest distance (km) to the ruptured plane}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s).}

\item{Fault_Type}{The indicator of fault type: 0 for strike slip & other non-reverse faulting;
1 for reverse}
}
\value{
An array of four values: (1) median spectral acceleration prediction; (2)
logarithmic standard deviation of spectral acceleration prediction; (3) logarithmic
standard deviation of within event residuals; and (4) logarithmic standard deviation of between
event residuals at the period of ip.
}
\description{
This is a subroutine of Idriss 2014
}
\references{
Idriss, I. M. (2014). An NGA-West2 Empirical Model for Estimating
the Horizontal Spectral Values Generated by Shallow Crustal Earthquakes.
Earthquake Spectra, 30(3), 1155-1177.
}
