% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{M_term}
\alias{M_term}
\title{The magnitude subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
M_term(M, V)
}
\description{
This subroutine function of BCHydro 2011 calculates magnitude scaling term.
DO NOT Run it. See bchydro_2011() for prediction.
}
