% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianLnYrefDataList.R
\docType{data}
\name{MedianLnYrefDataList}
\alias{MedianLnYrefDataList}
\title{The results of median GMM for reference conditions in CENA}
\format{
A list with five elements: MedianLnYrefDataTable - a 275 by 34 matrix saves all median results;
MedianLnYrefFrequencyLabelColumn - a 275 array saves the frequnecy name of MedianLnYrefDataTable matrix column;
MedianLnYrefFrequencyValueColumn - a 275 array saves the corresponding frequnecy values;
MedianLnYrefMomentColumn - a 275 array saves the corresponding moment magnitudes of MedianLnYrefDataTable matrix column;
MedianLnYrefMomentColumn - a 34 array saves the Rrup distances of MedianLnYrefDataTable matrix row.
}
\usage{
MedianLnYrefDataList
}
\description{
A list containing all results for the median GMM for reference conditions in CENA
}
\references{
Report PEER 2018/08 - Central and Eastern North America
Ground-Motion Characterization - NGA-East Final Report
Christine Goulet, Yousef Bozorgnia, Norman Abrahamson,
Nicolas Kuehn, Linda Al Atik, Robert Youngs, and Robert Graves
https://apps.peer.berkeley.edu/publications/peer_reports/reports_2018/NGA-East_1MainReport_FINAL.pdf
}
\keyword{datasets}
