% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nga_cena.R
\name{nga_cena_gmm}
\alias{nga_cena_gmm}
\title{The GMPE of NGA Central and Eastern North America (CENA) Model}
\usage{
nga_cena_gmm(M, T = 1000, Rrup, Vs30)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{T}{Period (sec); Use Period = -1 for PGV computation.
Use 1000 (by default) for output the array of Sa with pre-defined CENA periods: (-1,
0, 0.010, 0.020, 0.030, 0.040, 0.050, 0.075, 0.080, 0.100, 0.110, 0.112202, 0.114815,
0.11749, 0.12, 0.125893, 0.13, 0.134896, 0.14, 0.15, 0.2, 0.25, 0.3, 0.4,
0.5, 0.75, 0.8, 1.000, 1.500, 2.000, 3.000, 4.000, 5.000, 7.500, 10.000)}

\item{Rrup}{The closest distance (km) to the fault rupture}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s)}
}
\value{
A list of seven elements is returned: period - the input period of interest;
med - median spectral acceleration prediction (in g);
sigma - logarithmic standard deviation of spectral acceleration prediction;
med_ref - median spectral acceleration prediction on reference site condition;
Fs - the site amplification (in arithmetic); lnflin - the linear site amplification (in log);
lnfnl - the nonlinear site amplification (in log).
}
\description{
This function calculates Vs30-specific ground motion median values and standard deviations
}
\examples{
nga_cena_gmm(M = 6, T = 1000, Rrup = 30, Vs30 = 350)
}
\references{
Report PEER 2018/08 - Central and Eastern North America
Ground-Motion Characterization - NGA-East Final Report
Christine Goulet, Yousef Bozorgnia, Norman Abrahamson,
Nicolas Kuehn, Linda Al Atik, Robert Youngs, and Robert Graves
https://apps.peer.berkeley.edu/publications/peer_reports/reports_2018/NGA-East_1MainReport_FINAL.pdf

Hashash YMA, Ilhan O, Harmon JA, et al. Nonlinear site amplification model for ergodic seismic hazard analysis in Central and Eastern North America. Earthquake Spectra. 2020;36(1):69-86. doi:10.1177/8755293019878193

Stewart JP, Parker GA, Atkinson GM, Boore DM, Hashash YMA, Silva WJ. Ergodic site amplification model for central and eastern North America. Earthquake Spectra. 2020;36(1):42-68. doi:10.1177/8755293019878185
}
