% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nga_cena.R
\name{nga_cena_gmm_ref}
\alias{nga_cena_gmm_ref}
\title{The subroutine of CENA ground motion prediction on reference condition}
\usage{
nga_cena_gmm_ref(M, thisT, Rrup, SigmaType = "Ergodic")
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{thisT}{One single input period. -1 is for PGV, 0 is for PGA, and 0.01-10 for PSA at these oscillator period range.}

\item{Rrup}{The closest distance (km) to the fault rupture}

\item{SigmaType}{Type of Sigma: Ergodic or SingleStation. By default, it is Ergodic.}
}
\value{
a list of two elements, median GMM and standard deviation.
}
\description{
This is a subroutine of CENA GMM for reference condition
}
\references{
Report PEER 2018/08 - Central and Eastern North America
Ground-Motion Characterization - NGA-East Final Report
Christine Goulet, Yousef Bozorgnia, Norman Abrahamson,
Nicolas Kuehn, Linda Al Atik, Robert Youngs, and Robert Graves
https://apps.peer.berkeley.edu/publications/peer_reports/reports_2018/NGA-East_1MainReport_FINAL.pdf
}
