% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nga_cena.R
\name{nga_cena_gmmsiteamp}
\alias{nga_cena_gmmsiteamp}
\title{The subroutine of CENA site amplification function}
\usage{
nga_cena_gmmsiteamp(
  thisT,
  Vs30,
  PGAr,
  flin_coeffs = sea_2020_cena_flin_coeffs,
  fnl_coeffs = hea_2020_cena_fnl_coeffs
)
}
\arguments{
\item{thisT}{One single input period. -1 is for PGV, 0 is for PGA, and 0.01-10 for PSA at these oscillator period range.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s)}

\item{PGAr}{The PGA on reference rock condition (Vs30 = 3000 m/s)}

\item{flin_coeffs}{The coefficient for linear site amplification from Stewart et al 2020 CENA Flin coefficients.
You can use the internal saved data object, sea_2020_cena_flin_coeffs}

\item{fnl_coeffs}{The coefficient for nonlinear site amplification from Hashash et al 2020 CENA Fnl coefficients.
You can use the internal saved data object, hea_2020_cena_fnl_coeffs}
}
\description{
This is a subroutine of CENA GMM for site amplification
}
\references{
Hashash YMA, Ilhan O, Harmon JA, et al. Nonlinear site amplification model for ergodic seismic hazard analysis in Central and Eastern North America. Earthquake Spectra. 2020;36(1):69-86. doi:10.1177/8755293019878193

Stewart JP, Parker GA, Atkinson GM, Boore DM, Hashash YMA, Silva WJ. Ergodic site amplification model for central and eastern North America. Earthquake Spectra. 2020;36(1):42-68. doi:10.1177/8755293019878185
}
