% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pea_2020_coeffs.R
\docType{data}
\name{pea_2020_coeffs2}
\alias{pea_2020_coeffs2}
\title{The coefficients for Parker et al 2020}
\format{An object of class \code{data.frame} with 26 rows and 46 columns.}
\source{
\url{https://apps.peer.berkeley.edu/publications/peer_reports/reports_2020/PSHAB_Electronic_Supplement.zip}
}
\usage{
pea_2020_coeffs2
}
\description{
The Coefficients for Slab
}
\keyword{datasets}
