% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pea_2020_ngasub.R
\name{pea_2020_ngasub}
\alias{pea_2020_ngasub}
\title{The GMPE of Parker at al 2020}
\usage{
pea_2020_ngasub(
  M,
  T = 1000,
  Rrup,
  Vs30,
  event.type,
  region = "Global",
  saturation.region = "Global",
  Z2.5 = "default",
  basin = 0,
  hypocentral.depth,
  Ts = NA,
  Reg = "Global",
  coefficients1 = pea_2020_coeffs1,
  coefficients2 = pea_2020_coeffs2,
  coefficients3 = pea_2020_coeffs3,
  coefficients4_IF = pea_2020_coeffs4_Interf,
  coefficients4_Slab = pea_2020_coeffs4_Slab
)
}
\arguments{
\item{M}{Moment magnitude, a numeric value}

\item{T}{Period (sec); Use Period = -1 for PGV computation.
Use 1000 (by default) for output the array of Sa with original Pea et al pre-defined periods}

\item{Rrup}{Closest distance (km) to the ruptured plane}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s)}

\item{event.type}{The indicator of subduction type: 0 for interface; 1 for slab}

\item{region}{A string that corresponds to options in the DatabaseRegion column of the flatfile,
plus global. Must be a string in any of: "Global", "Alaska", "Cascadia", "CAM", "Japan", "SA" or "Taiwan".
If no matches, default will be global model.}

\item{saturation.region}{A string that corresponds to regions defined by C. Ji and R. Archuleta (2018):
"Global", "Aleutian","Alaska","Cascadia","CAM_S", "CAM_N", "Japan_Pac","Japan_Phi",
"SA_N","SA_S", "Taiwan_W","Taiwan_E". If unknown, use "Global".}

\item{Z2.5}{The basin depth to the layer where the shear wave is 2.5 km/s.
Only used if DatabaseRegion == "Japan" or "Cascadia".
Can also specify "default" to get no basin term (e.g. delta_Z2.5 = 0)}

\item{basin}{Flag of baisn type. It is only used if DatabaseRegion == "Cascadia".
Value can be 0, 1, or 2,
where 0 == having an estimate of Z2.5 outside mapped basin, 1 == Seattle basin,
and 2 == other mapped basin (Tacoma, Everett, Georgia, etc.)}

\item{hypocentral.depth}{The depth of hypocenter (km).
To use Ztor value to estimate hypocentral depth, see Ch. 4.3.3/Eqs. 4.16 & 4.17
of Parker et al. (2020) PEER report}

\item{Ts}{A single period ranges from 0.01-10s for epistemic uncertainty.
Ts = NA will take the same periods as T.}

\item{Reg}{A string to specify region for epistemic uncertainty. Options include
"Global", "Alaska", "Aleutian", "Cascadia", "CAM_N", "CAM_S", "Japan_Pac", "Japan_Phi",
"SA_N", "SA_S", "Taiwan".}

\item{coefficients1}{The model coefficients for interface. Use internal saved data, no need to specify.}

\item{coefficients2}{The model coefficients for slab. Use internal saved data, no need to specify.}

\item{coefficients3}{The model coefficients for aleatory. Use internal saved data, no need to specify.}

\item{coefficients4_IF}{The model coefficients for epistemic for interface. Use internal saved data, no need to specify.}

\item{coefficients4_Slab}{The model coefficients for epistemic for slab Use internal saved data, no need to specify.}
}
\value{
A dataframe of 15 variables. 1. period, the same as input; 2. Vs30, the same as input;
3. Rrup, the same as input; 4. Tau, even-to-event Std; 5. Phi_total, total within-event Std;
6. Phi_SS, single station Std; 7. PhiS2S, site-to-site Std;
8. Phi_partitioned, remaining Std - Phi_total excludes Phi_SS and PhiS2S;
9. Sigma_total, total Std; 10. Sigma_partitioned, Sigma_total excludes Tau and Phi_total;
11. Mu, median prediction (g - PSA/PGA, cm/s - PGV); 12. Flin, linear amplification;
13. Fnl, nonlinear amplification; 14. Rock, median prediction on reference rock site (g - PSA/PGA,
cm/s - PGV); 15. Epi, epistemic uncertainty
}
\description{
This function calculates the ground motion median values and standard deviations
}
\examples{
pea_2020_ngasub(M = 7.8, T = 1000, Rrup = 50, Vs30 = 350, event.type = 1,
hypocentral.depth = 45)

pea_2020_ngasub(M = 8.5, T = c(0.035, 6), Rrup = 50, Vs30 = 450, event.type = 0,
region = "Japan", saturation.region = "Japan_Pac", Z2.5 = "default", hypocentral.depth = 75,
Ts = NA, Reg = "Japan_Pac")
}
\references{
Parker, GA, JP Stewart, DM Boore, GM Atkinson, B Hassani (2020).
NGA-Subduction global ground-motion models with regional adjustment factors,
PEER Report 2020/03, Pacific Earthquake Engineering Research Center,
UC Berkeley (Center Headquarters), 131 pages.
}
