% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchydro_2011_subduction.R
\name{Site_term}
\alias{Site_term}
\title{The site effect subroutine function of BCHydro 2011 for subduction earthquakes}
\usage{
Site_term(Vs30, PGA_rock, V)
}
\description{
This subroutine function of BCHydro 2011 calculates site response term.
DO NOT Run it. See bchydro_2011() for prediction.
}
